+++ user-api/user-infrastructure/src/main/java/kr/flab/ottsharing/user/infrastructure/config/UserRepositoryConfig.java
@@ -1,15 +1,15 @@
 package kr.flab.ottsharing.user.infrastructure.config;
 
 import kr.flab.ottsharing.user.domain.repository.UserV1Repository;
-import kr.flab.ottsharing.user.infrastructure.config.repository.jpa.UserV1JpaRepository;
-import kr.flab.ottsharing.user.infrastructure.config.repository.UserV1RepositoryAdaptor;
+import kr.flab.ottsharing.user.infrastructure.repository.jpa.UserV1JpaRepository;
+import kr.flab.ottsharing.user.infrastructure.repository.UserV1RepositoryAdaptor;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @Configuration
-@EnableJpaRepositories(basePackages = "kr.flab.ottsharing.user.infrastructure.config.repository.jpa")
+@EnableJpaRepositories(basePackages = "kr.flab.ottsharing.user.infrastructure.repository.jpa")
 @EntityScan(basePackages = "kr.flab.ottsharing.user.domain")
 public class UserRepositoryConfig {
