--- user-api/user-infrastructure/src/main/java/kr/flab/ottsharing/user/infrastructure/config/repository/jpa/EmailTokenJpaRepository.java
@@ -1,10 +1,0 @@
-package kr.flab.ottsharing.user.infrastructure.config.repository.jpa;
-
-import kr.flab.ottsharing.user.domain.EmailToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EmailTokenJpaRepository extends JpaRepository<EmailToken, Long> {
-}
-
