+++ user-api/user-infrastructure/src/main/java/kr/flab/ottsharing/user/infrastructure/repository/jpa/EmailValidCheckTokenJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmailTokenJpaRepository extends JpaRepository<EmailValiedCheckToken, Long> {
+public interface EmailValidCheckTokenJpaRepository extends JpaRepository<EmailValiedCheckToken, Long> {
 }
